requirements.txt
mecab-python3==0.996.5
PyYAML==5.3.1
dill==0.3.1.1
neologdn==0.4
iscord.py==1.3.3